automatically add a `sent_at` property to the identify command's payload
I found the tests hard to read, so I completely refactored them with the addition of several helper functions to handle boilerplate.

I also removed the tests that validate the ability to override the `active` property of the context. That functionality is part of the base class, `WithContext`, having just written unit tests for that a couple commits ago, I felt these were redundant.